Factor out _format_unit_list
