+++ astropy/units/format/base.py
@@ -1,4 +1,5 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
+from . import utils
 
 
 class Base:
@@ -7,6 +8,7 @@ class Base:
     """
 
     registry = {}
+    _space = " "
 
     def __new__(cls, *args, **kwargs):
         # This __new__ is to make it clear that there is no reason to
@@ -33,6 +35,28 @@ class Base:
         Convert a string to a unit object.
         """
         raise NotImplementedError(f"Can not parse with {cls.__name__} format")
+
+    @classmethod
+    def _format_superscript(cls, number):
+        return f"^{number}"
+
+    @classmethod
+    def _format_unit_power(cls, unit, power=1):
+        """Format the unit for this format class raised to the given power.
+
+        This is overridden in Latex where the name of the unit can depend on the power
+        (e.g., for degrees).
+        """
+        name = cls._get_unit_name(unit)
+        if power != 1:
+            name += cls._format_superscript(utils.format_power(power))
+        return name
+
+    @classmethod
+    def _format_unit_list(cls, units):
+        return cls._space.join(
+            cls._format_unit_power(base_, power) for base_, power in units
+        )
 
     @classmethod
     def to_string(cls, u):
