+++ astropy/units/format/ogip.py
@@ -391,20 +391,8 @@ class OGIP(generic.Generic):
         return name
 
     @classmethod
-    def _format_unit_list(cls, units):
-        out = []
-        units.sort(key=lambda x: cls._get_unit_name(x[0]).lower())
-
-        for base, power in units:
-            if power == 1:
-                out.append(cls._get_unit_name(base))
-            else:
-                power = utils.format_power(power)
-                if "/" in power:
-                    out.append(f"{cls._get_unit_name(base)}**({power})")
-                else:
-                    out.append(f"{cls._get_unit_name(base)}**{power}")
-        return " ".join(out)
+    def _format_superscript(cls, number):
+        return f"**({number})" if "/" in number else f"**{number}"
 
     @classmethod
     def to_string(cls, unit):
